Adjust: Move logic that verify that user is logged in from dashboard view to the main view
